Add the Dockerfile required to create the docker image
A few changes made to the docker-compose file, now the app should run as a service from the docker-compose. Using docker containers to run MinIO and to run the java application causes some issues with the generated pdf url from MinioClient, I added a temporary fix for setting the url method in MinioService.